Hook for å hente og lagre status om delt tiltak med bruker
