+++ frontend/mulighetsrommet-veileder-flate/src/core/api/query-keys.ts
@@ -3,4 +3,5 @@ export const enum QueryKeys {
   Brukerdata = 'brukerdata',
   Veilederdata = 'veilederdata',
   Historikk = 'historikk',
+  DeltMedBrukerStatus = 'deltMedBrukerStatus',
 }
