+++ mulighetsrommet-api/src/main/resources/web/openapi.yml
@@ -309,7 +309,7 @@ paths:
             type: string
           description: "Veileders ident"
         - in: query
-          name: "tiltakstnummer"
+          name: "tiltaksnummer"
           schema:
             type: string
           description: "Tiltaksnummer"
